requirements.txt
kafka-python==2.1.5
pandas==1.5.3
numpy==1.23.5
requests
apache-airflow==2.10.5
apache-airflow-providers-openlineage>=1.8.0
psycopg2-binary
python-dotenv


